+++ js/api.js
@@ -1,0 +1,27 @@
+/**
+ * @copyright   2010-2015, The Titon Project
+ * @license     http://opensource.org/licenses/BSD-3-Clause
+ * @link        http://titon.io
+ */
+
+define([
+    './events/clickout',
+    './events/horizontal-resize',
+    './events/swipe',
+    './events/vertical-resize',
+    './extensions/aria',
+    './extensions/bound',
+    './extensions/cache',
+    './extensions/conceal',
+    './extensions/debounce',
+    './extensions/position-to',
+    './extensions/reveal',
+    './extensions/shown-selector',
+    './extensions/throttle',
+    './extensions/to-string',
+    './extensions/toolkit',
+    './extensions/transitionend'
+], function() {
+    // An empty file that includes all event and extension dependencies.
+    // This should make these always available, even if the build is customized.
+});