+++ js/components/modal.js
@@ -58,10 +58,13 @@ Toolkit.Modal = Toolkit.Component.extend({
             //['clickout', 'element', 'onHide'],
             //['clickout', 'document', 'onHide', '{selector}'],
             ['click', 'document', 'onShow', '{selector}'],
-            ['click', 'element', 'onHide'],
             ['click', 'element', 'hide', this.ns('close')],
             ['click', 'element', 'onSubmit', this.ns('submit')]
         ]);
+
+        if (options.clickout) {
+            this.addEvent('click', 'element', 'onHide');
+        }
 
         this.initialize();
     },
@@ -225,10 +228,10 @@ Toolkit.Modal = Toolkit.Component.extend({
 
 }, {
     animation: 'fade',
-    ajax: true,
     blackout: true,
     fullScreen: false,
     stopScroll: true,
+    clickout: true,
     getContent: 'data-modal',
     template: '<div class="' + vendor + 'modal">' +
         '<div class="' + vendor + 'modal-outer">' +